fixed a bug with signed int32 -> unsigned int32 -> signed int32 and so on...

+++ src/LoadCell.cpp
@@ -43,6 +43,6 @@ double LoadCell::read(uint8_t i) {
   return loadcell_.read_average(i);
 }
 
-uint32_t LoadCell::quick_read() {
-  return loadcell_.read();
+long LoadCell::quick_read() {
+  return static_cast<int32_t>(loadcell_.read());
 }
